requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
pydantic==2.10.4
httpx==0.27.0
python-dotenv==1.0.1
